requirements.txt
cloudpickle==2.0.0
cycler==0.11.0
fonttools==4.28.1
gym==0.21.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
